login check, doAdd에 writer정보저장

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.koreaIT.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,20 +27,25 @@ public class UsrArticleController {
 	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
 		
-		if (Util.empty(title)) { 
+		if (httpSession.getAttribute("loginMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+		
+		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
-		if (Util.empty(body)) { 
+		
+		if (Util.empty(body)) {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 		
-		articleService.writeArticle(title, body);
+		articleService.writeArticle(title, body, (int) httpSession.getAttribute("loginMemberId"));
 		
 		int id = articleService.getLastInsertId();
 		
-		return ResultData.from("S-1", Util.f("%d번 게시물이 생성되었습니다.", id), articleService.getArticleById(id));
+		return ResultData.from("S-1", Util.f("%d번 게시물이 생성되었습니다", id), articleService.getArticleById(id));
 	}
 	
 	@RequestMapping("/usr/article/getArticle")
@@ -48,45 +55,54 @@ public class UsrArticleController {
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
-		return ResultData.from("S-1", Util.f("%d번 게시물입니다.", id), article);
+		return ResultData.from("S-1", Util.f("%d번 게시물 입니다", id), article);
 	}
 	
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	public List<Article> getArticles() {
-		return articleService.getArticles();
+	public ResultData<List<Article>> getArticles() {
+		
+		return ResultData.from("S-1", "게시물 리스트", articleService.getArticles());
 	}
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public String doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		if (httpSession.getAttribute("loginMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return id + "번 게시물은 존재하지 않습니다.";
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.modifyArticle(id, title, body);
 		
-		return id + "번 게시물을 수정했습니다";
+		return ResultData.from("S-1", Util.f("%d번 게시물을 수정했습니다", id));
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public String doDelete(int id) {
+	public ResultData doDelete(HttpSession httpSession, int id) {
+		
+		if (httpSession.getAttribute("loginMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return id + "번 게시물은 존재하지 않습니다.";
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.deleteArticle(id);
 		
-		return id + "번 게시물을 삭제했습니다";
+		return ResultData.from("S-1", Util.f("%d번 게시물을 삭제했습니다", id));
 	}
 }
